requirements.txt
fire==0.3.1
GitPython==3.1.27
packaging==21.3
